feat: delete files with storage file remotion

+++ domain/use-cases/property-image.cases.ts
@@ -1,3 +1,4 @@
+import { clearPathFiles } from "@/lib/utils";
 import { PropertyImageEntity } from "../entities/property-image.entity";
 import { PropertyImagePort } from "../ports/property-image.port"
 
@@ -26,8 +27,29 @@ export class PropertyImageUseCases {
     return this.repository.update(exist.id ?? '', input)
   }
 
-  delete(id: string) {
+  async delete(id: string) {
+    const exist = await this.repository.getById(id);
+
+    if (!exist) {
+      throw new Error("El recurso no fue encontrado");
+    }
+    
+    if (exist.public_url) {
+      await this.deleteFile(exist.public_url)
+    }
     return this.repository.delete(id)
+  }
+
+  async deleteFile(public_url: string) {
+    // remove params from url to get the clear path
+    const clearUrl = clearPathFiles(public_url)
+    const exist = await this.repository.existPath(clearUrl)
+
+    if (!exist) {
+      throw new Error("El archivo no existe")
+    }
+    
+    return this.repository.deleteFile(clearUrl)
   }
 
   async uploadFile(propertyId: string, name: string, image: File): Promise<string> {
